+++ packages/model-serving/src/components/deploymentWizard/fields/AvailableAiAssetsFields.tsx
@@ -1,8 +1,11 @@
 import React from 'react';
-import { Checkbox, TextInput, StackItem, Stack, FormGroup } from '@patternfly/react-core';
+import { Checkbox, TextInput, StackItem, Stack, FormGroup, Popover } from '@patternfly/react-core';
 import { z } from 'zod';
 import { ServingRuntimeModelType } from '@odh-dashboard/internal/types';
+import DashboardPopupIconButton from '@odh-dashboard/internal/concepts/dashboard/DashboardPopupIconButton';
+import { OutlinedQuestionCircleIcon } from '@patternfly/react-icons';
 import { ModelTypeFieldData } from './ModelTypeSelectField';
+import { UseModelDeploymentWizardState } from '../useDeploymentWizard';
 
 export type AvailableAiAssetsFieldsData = {
   saveAsAiAsset: boolean;
@@ -52,11 +55,13 @@ export const useAvailableAiAssetsFields = (
 type AvailableAiAssetsFieldsComponentProps = {
   data: AvailableAiAssetsFieldsData;
   setData: (data: AvailableAiAssetsFieldsData) => void;
+  wizardData: UseModelDeploymentWizardState;
 };
 
 export const AvailableAiAssetsFieldsComponent: React.FC<AvailableAiAssetsFieldsComponentProps> = ({
   data,
   setData,
+  wizardData,
 }) => {
   const resetAiAssetData = React.useCallback(
     (save: boolean) => {
@@ -67,30 +72,50 @@ export const AvailableAiAssetsFieldsComponent: React.FC<AvailableAiAssetsFieldsC
     },
     [setData],
   );
+  const showSaveAsAiAsset = React.useMemo(() => {
+    if (wizardData.state.modelType.data === ServingRuntimeModelType.GENERATIVE) return true;
+    return false;
+  }, [wizardData.state.modelType.data]);
 
-  return (
-    <Stack hasGutter>
-      <StackItem>
-        <Checkbox
-          id="save-as-ai-asset-checkbox"
-          data-testid="save-as-ai-asset-checkbox"
-          label="Make this deployment available as an AI asset"
-          isChecked={data.saveAsAiAsset}
-          onChange={(_, checked) => resetAiAssetData(checked)}
-        />
-      </StackItem>
-      {data.saveAsAiAsset && (
-        <StackItem>
-          <FormGroup label="Use case">
-            <TextInput
-              id="use-case-input"
-              data-testid="use-case-input"
-              value={data.useCase}
-              onChange={(_, value) => setData({ ...data, useCase: value })}
+  return showSaveAsAiAsset ? (
+    <StackItem>
+      <FormGroup
+        label="AI Asset"
+        data-testid="ai-asset-section"
+        fieldId="ai-asset"
+        labelHelp={
+          <Popover bodyContent="POPOVER BODY CONTENT" headerContent="POPOVER HEADER CONTENT">
+            <DashboardPopupIconButton
+              icon={<OutlinedQuestionCircleIcon />}
+              aria-label="More info"
             />
-          </FormGroup>
-        </StackItem>
-      )}
-    </Stack>
-  );
+          </Popover>
+        }
+      >
+        <Stack hasGutter>
+          <StackItem>
+            <Checkbox
+              id="save-as-ai-asset-checkbox"
+              data-testid="save-as-ai-asset-checkbox"
+              label="Make this deployment available as an AI asset"
+              isChecked={data.saveAsAiAsset}
+              onChange={(_, checked) => resetAiAssetData(checked)}
+            />
+          </StackItem>
+          {data.saveAsAiAsset && (
+            <StackItem>
+              <FormGroup label="Use case">
+                <TextInput
+                  id="use-case-input"
+                  data-testid="use-case-input"
+                  value={data.useCase}
+                  onChange={(_, value) => setData({ ...data, useCase: value })}
+                />
+              </FormGroup>
+            </StackItem>
+          )}
+        </Stack>
+      </FormGroup>
+    </StackItem>
+  ) : null;
 };
